feat: create data listing project structure
As I am studying the Tailwind CSS structure, it would be great to use it
in different projects, such as this data listing project created based
a Rocketseat's video called "Nivo 2", designed during the "React na
Prática" event, one of their free programing events during the year.

This commit introduces the whole project structure which is going to be
used during my study. I am going to make frequent updates to this
repository, that is why issues may be detected. Fell free to report!